requirements.txt
python-telegram-bot[job-queue]==22.0
aiosqlite>=0.19.0
httpx>=0.25.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
pytest>=7.0
pytest-asyncio>=0.23.0
